build.zig: run prover on installed binaries
Until now, the install step for the guest stf binaries was run _after_
the run step, which means that the prover was running outdated binaries.
Make the cli executable step depend on the install step of the gues stf
binaries. Since the run step depends on the cli binary being built, it
will use the latest binaries when it starts.

Signed-off-by: Guillaume Ballet <3272758+gballet@users.noreply.github.com>