zisk: force data at 0xa0020000
If zig produces a binary without data/bss sections, the computation
for the stack top will end up in the code. This wouldn't work. This
fix prevents that situation by ensuring the stack starts exactly at
the address it's supposed to be by forcing the address. Most of the
time, this is a no-op, but in the case described above, the correct
behavior is forced.

Signed-off-by: Guillaume Ballet <3272758+gballet@users.noreply.github.com>