zisk: fix allocator
zisk binaries have to be built in pie mode, which means that
all the values defined in the linker script are not available
without a relocation. This is most inconvenient, as we don't
want to maintain our custom loader. Unless we change our minds
down the road (or find a fix for the PIE version, or force a
different addressing mode), this forces the symbols to be absolute
and then all other missing values are inferred doing pointer math.

Signed-off-by: Guillaume Ballet <3272758+gballet@users.noreply.github.com>