tentative CI fix
The run step wasn't dependent on the cli executable being built, so
installation would not occur. That is probably not the best way to
do it, since it would be better if it depended directly on the
installer. I am still hoping to rewrite that part in a more streamlined
way, so I would rather commit this now and correct it all, rather than
optimize something I know won't stick too long.